Heater Reset 2 Hour Delay
Occasionally the hardware switch controlling the heater would become stuck in the on or off position, so a reset function was introduced that power cycled the switch on/off if the water temperature was greater than or less than 2 degrees of the thermostat setting. This resolved the stuck switch condition, but it introduced a new problem of frequent resetting when changing the thermostat setting more than a couple of degrees. A new lastUpdated variable now keeps track of when the thermostat was last changed. The controller will only reset the switch if it has been more than 2 hours since the thermostat was last modified and the water temp is greater than or less than 1.5 degrees of the set temp.